requirements.txt
python-docx
pandas
openpyxl
requests
markdownify
